requirements.txt
# -- Usage -- 
# $ pip install -r requirements.txt

discord.py == 2.1.0